@@ -1,6 +1,6 @@
 {
-  "name": "@aca-1/a2-widgets",
-  "version": "0.8.0",
+  "name": "@acaprojects/a2-widgets",
+  "version": "0.8.5",
   "description": "Angular 2 Widgets Library",
   "main": "index.js",
   "scripts": {
@@ -44,13 +44,13 @@
     "zone.js": "^0.8.5"
   },
   "devDependencies": {
-    "@types/chai": "^3.4.34",
+    "@types/chai": "^4.0.0",
     "@types/hammerjs": "^2.0.33",
-    "@types/jasmine": "2.5.47",
+    "@types/jasmine": "2.5.51",
     "@types/moment": "^2.13.0",
     "@types/node": "^7.0.11",
     "@types/protractor": "^4.0.0",
-    "@types/selenium-webdriver": "3.0.3",
+    "@types/selenium-webdriver": "3.0.4",
     "@types/source-map": "^0.5.0",
     "@types/uglify-js": "^2.0.27",
     "@types/webpack": "^2.2.12",
@@ -64,7 +64,7 @@
     "del": "^2.2.2",
     "dotenv": "^4.0.0",
     "gulp": "^3.9.1",
-    "gulp-autoprefixer": "^3.1.1",
+    "gulp-autoprefixer": "^4.0.0",
     "gulp-clean": "^0.3.2",
     "gulp-inject-file": "0.0.18",
     "gulp-replace": "^0.5.4",
@@ -89,8 +89,8 @@
     "ts-loader": "^2.0.3",
     "tslint": "^5.0.0",
     "tslint-loader": "^3.4.3",
-    "typescript": "2.3.2",
-    "webpack": "2.5.1",
+    "typescript": "2.3.4",
+    "webpack": "2.6.1",
     "webpack-dev-middleware": "^1.6.1",
     "webpack-dev-server": "2.4.5",
     "webpack-md5-hash": "^0.0.5",
